ir.model: don't re-create records that were deleted by the user
WARNING: behaviour change on module-updates

See, a few commits ago we introduced the concept of "res_id == 0" for
deleted records. This way we know that the user has explicitly deleted
something.
If a module was updated, it would try to update/create _all_ missing
records, including the deleted. Now, it will just skip these.