rml2pdf: implement the 'storyPlace' flowable tag
As in RML User's Guide p.6.4 , storyPlace can render out-of-band
flowables.
In our case, it is trivial for extra frames that need to take part
of our story, but NOT be treated like the main (story) frame. Say,
we need a little "sum" box at the bottom of an invoice, which will
only take the total sum at the end of (invoice) story.