orm: remove old code from read_group()
So far, read_group() almost works like in 6.0, /except/ for sorting,
that is not yet implemented.
It also differs in the results, because F3 can now provide values for
*all* fields, not only int/float ones. As this is an improvement, we
allow the new API to do that.

F3 mode will also need a little more work, as it has hardly had any
testing so far.