netsvc: accumulate function calls through setAlarm()
setAlarm() is a way to defer actions, to run at a given time, using
a separate cursor (possibly).
But, sometimes, we need to merge multiple calls of those actions,
when the called function may operate with many records at once.

Suppose we have this action:
   def fix_my_records(db_name, uid, record_ids):
        cr = ... cursor()
        for rec_id in record_ids:
              do_fix_that(rec_id)

	return True

If we have a calling function that will queue
    `setAlarm(fix_my_records, cr.dbname, [rec.id],...)`
that may be sub-optimal, because it would trigger `fix_my_records()`
multiple times.

Now, we can set:
    fix_my_records._cumulative = 2
meaning that the 3rd argument of the function can be accumulated as
a list.
Hopefully, this will boost performance through reusing a single call
and thus the same cursor.