sql_db: post-commit triggers for cursors
After a cursor has been committed, allow some custom triggers to run,
depending on the operations performed. Respect the transactional logic
of SQL-based operations, outside the database scope.

This is a big feature, queued since 2010. Some ORM operations, like
ones involving files on the server's filesystem, perform actions out
of the SQL-scope, but need to be synchronized with what is written to
the database. Actions need to be run /after/ the cursor is committed,
or in the event that it is rolled back.