netsvc: support fully-cumulative functions in Agent
Cumulative functions are ones that need only run once, instead of
multiple times. So far, we had expected that one of their arguments
would be the list of IDs (or so) to append to while accumulating.

Now, with `cumulative_on=True` we can have functions that need not
append to any argument. Like:
    def process_all_spam(cr, uid, context):
        ...
Which would have the same result regardless.