osv: deferred ORM call implementation
So far, all our ORM methods would run synchronously, under the same
process, db cursor etc.
But it's becoming a pattern that we need to run methods in an other
thread, after our first one has finished.
agent.setAlarm() would do that, but we need to wrap the ORM method
in a cursor environment.

`defer()` does that, easily. It transforms a classic method into a
deferred, taking care of cursors and exception logging.

Note that it is also transaction-sensitive, meaning that a rollback
on the caller will cancel the deferred call.