res.users: "sudo()" call to methods as another user
This gives an ORM entry to call a method as an arbitrary user,
without actually authenticating as that uid. Limited to admin,
of course.

There is many situations when we want to check if some user has
the rights to perform some operation, but logging-in as him/her
is not easy. `sudo()` is a way for the admin to quickly test
such a case, from a remote client.