fields/relational: allow '_vptr' writes to x2many fields
At x2many fields, verb `(0,0,{})` was previously allowed to specify
`_vptr` and instead create a virtual object of some child class.
Now, we expand that to verb `(1,ID,{})` in order to write fields in
these virtual objects.

This helps multiple writes to sub-records of our main record, in a
single RPC call.