netsvc: protect cumulative function algo. from exceptions
With cumulative functions, we acquire the Agent lock, compare
against existing requests and then either merge or queue as a
separate request. However, if that comparison would fail with
an exception (the case when different types of arguments were
supplied), the lock would go stale and no request placed.

This got even worse, as deferred calls were part of db cursor
`on_commit()` post-handler. An exception from the netsvc code
would break the commit and leave the cursor open.