tools/server_types: add (in)equality operator to server_bool type
That type is artificial, cannot inherit built-in `bool` singletons.
So, the comparison "a == b" where both a and b are `server_bool`s
would fail because `__cmp__` of 'a' would take the boolean method.

This broke the cumulative functions in `netsvc.Agent`, in particular,
because they try to compare "contexts" containing server bools.