orm: implement "merge" for models with `inherits`
Eg. 'product.product' and 'project.project' models, they do have a
parent, but are also high in demand for records merging.

The algorithm, here, has to merge the parent, first, and then do
the rest of the _columns. For the parent, instead of a numeric ID,
we pass a dictionary of merge values.