openerp-server: adapt to "systemd", transparently
When launched as a systemd unit, the F3 server shall be configured
with "Type=notify" . Then, systemd will expect it to advertise its
state to the notification socket.
This is much better than SysV scripts, where we used to know only
that the server process was running, but not when/if it would be
ready to serve any requests.

This API is really optional/transparent: it will detect systemd from
the os.environment and just /try/ to emit the notifications. For
older systems, it will just remain silent.