osv/orm: create_or_update() method in core ORM
There is many cases where we want to push data from a sync source,
doing either update of an existing record or creating a new one.
In libclient, this would take two (or more) calls, an expensive
roundtrip.
Instead, put that simple algorithm in the core, let records be
pushed in a single RPC call.

Pending: a multi-record version of this call.