fields/special: don't store "server" values in struct field
We have a few "server" types (eg. server_int), which are supposed
to only work inside server function calls. We don't want them to
be stored in struct fields, either.

For example `context` may contain `process=server_bool(False)` to
suppress processing. This is not allowed to replay if we serialize
the context into a struct.