netsvc: armor Agent against errors with lock held
If `Agent.runner()` fails, we may never learn the exception, or
even have a dead thread with the _lock held. In such a case, any
attempt to touch the Agent would result in a pythonic deadlock.

`Agent.cancel()` shall also work regardless of the lock, so that
ir.cron calls won't block with some Postgres locking in place.