base/ir: fix "model overview" report advertising SXW file as RML
This would break the 'ir.actions.report.xml' view, badly, in the
GTK client, as the non-XML content would be passed as a string
that XML-RPC Expat parser cannot handle.

Thus, we need both to fix the data definition and put a workaround
for older databases which still hold the wrong "rml_content" path.