http_server: resolve client IP in proxy-aware way
After exposing our ERP to proxy servers, it is now useful to resolve
the real IP of the connected HTTP client.

In case of a forwarded service (through web proxy), we have to
analyze the "X-Forwarded-For" header.

Still, you need to explicitly define the known proxies in

[httpd]
trusted_proxies = 1.2.3.4[, IP-of-proxy2 ...]

so that we only trust that header from real proxies.

Note that the "client-pit" will also receive the resolved IP from
now on (implementation would tolerate a string instead of a tuple).