orm: avoid returning non-requested fields in read(), in wrong form
During a `read()`, internal cache is populated with more fields,
especially those needed to compute function fields of the model.
But, that cache contains the _classic_write form (ie. just id of
many2one fields). Then, non-requested fields are NOT passed to
post_process, causing wrong form of the data to be returned.

This broke the f3 awebui, in particular.