orm: shortcut `_ids` attribute for browse lists
We have been using `only_ids(browse_list_ids)` so far, to get the
pure list of ids. But this function may not be available in the
sandbox, nor convenient.

Instead, be able to write `brose_list._ids` to get that.

Note the underscore, which avoids conflict with any existing method.